Develop #9

+++ src/api/index.js
@@ -1,8 +1,10 @@
 const Router = require('koa-router');
 const model = require('../database/models');
 const token = require('../lib/token');
+const fetch = require('node-fetch');
 const crypto = require('crypto');
 const uuidv4 = require('uuid/v4');
+require('dotenv').config();
 
 const api = new Router();
 
@@ -34,6 +36,13 @@ api.get('/', async (ctx, next) => {
 // res: 성공 - OK(200) / 에러 - Error message(500)
 // Front-end에서 검증(비어 있지 않은 값, 타입 일치, 중복되지 않은 Id)된 값을 전달하므로 실패할 수 없음
 api.post('/users', async (ctx, next) => {
+    if(ctx.request.headers.referer !== process.env.SIGNUP_REFERER) {
+        console.log("[User]Create Failed: Unkown referer");
+        ctx.body = "Bad request.";
+        ctx.status = 400;
+        return;
+    }
+
     const { userId, userPw, name, engName } = ctx.request.body;
 
     // // userId, userPw, name, engName이 undefined인 경우는 front-end에서 소거, type은 string으로 보장
@@ -236,7 +245,14 @@ api.delete('/users', async (ctx, next) => {
 // req: userId(기존 유저 Id/string), userPw(기존 유저 Pw/string)
 // res: 성공 - OK(200) + Access token(+) / 실패 - Fail message(401) / 에러 - Error message(500)
 api.post('/auth', async (ctx, next) => {
-    const { userId, userPw, codeData } = ctx.request.body;
+    if(ctx.request.headers.referer !== process.env.LOGIN_REFERER) {
+        console.log("[Auth]Create Failed: Unkown referer");
+        ctx.body = "Bad request.";
+        ctx.status = 400;
+        return;
+    }
+
+    const { userId, userPw, codeData, kakaoToken } = ctx.request.body;
 
     // // userId, userPw가 undefined인 경우는 front-end에서 소거, type은 string으로 보장
     // // client 연산의 임시 코드
@@ -272,8 +288,52 @@ api.post('/auth', async (ctx, next) => {
     // });
 
     // if(duplicate) {
+        // 카카오 로그인 시도
+        if(kakaoToken !== "") {
+            let userId = "";
+            let nickName = "";
+
+            await fetch("https://kapi.kakao.com/v2/user/me", { headers: { "Authorization": "Bearer " + kakaoToken }})
+            .then(res => res.json()).then(result => {
+                // 누군가가 비정상적인 접근 시도를 하는 경우
+                if(result.kakao_account === undefined) {
+                    console.log("[Auth]Create Failed: Invalid Kakao Token");
+                    ctx.body = "Invalid kakao token.";
+                    ctx.status = 401;
+                }
+                else {
+                    userId = result.id;
+                    nickName = result.kakao_account.profile.nickname;
+                }
+            }).catch(err => {
+                console.log(err);
+                ctx.status = 500;
+            });
+
+            if(nickName !== "") {
+                const temp = crypto.randomBytes(64).toString('base64');
+                const salt = crypto.randomBytes(64).toString('base64');
+                const tempPw = crypto.pbkdf2Sync(temp, salt, 10000, 128, 'sha512').toString('base64');
+
+                await model.sequelize.models.Users.findOrCreate({
+                    where: { userId: userId },
+                    defaults: {
+                        userId: userId, userPw: tempPw, salt: salt, name: nickName, engName: nickName
+                    }
+                }).then(() => {
+                    console.log("[Auth]Create Success: Token Created");
+                    const accessToken = token.generateToken({ userId: nickName });
+
+                    ctx.cookies.set("accessToken", accessToken, { maxAge: 1000 * 60 * 60 * 21, sameSite: "Strict", overwrite: true });
+                    ctx.status = 200;
+                }).catch(err => {
+                    console.log(err);
+                    ctx.status = 500;
+                });
+            }
+        }
         // 일반 로그인 시도
-        if(userId !== "" && userPw !== "") {
+        else if(userId !== "" && userPw !== "") {
             await model.sequelize.models.Users.findOne({
                 where: { userId: userId }
             }).then(result => {
@@ -288,7 +348,7 @@ api.post('/auth', async (ctx, next) => {
                     console.log("[Auth]Create Success: Token Created");
                     const accessToken = token.generateToken({ userId: userId });
     
-                    ctx.cookies.set("accessToken", accessToken, { httpOnly: false, maxAge: 1000 * 60 * 60 * 21 });
+                    ctx.cookies.set("accessToken", accessToken, { maxAge: 1000 * 60 * 60 * 21, sameSite: "Strict", overwrite: true });
                     ctx.status = 200;
                 }
             }).catch(err => {
@@ -297,7 +357,7 @@ api.post('/auth', async (ctx, next) => {
             });
         }
         // QR 로그인 시도
-        else {
+        else if(codeData !== "") {
             await model.sequelize.models.QRCodes.findOne({
                 where: { codeData: codeData }
             }).then(async result => {
@@ -314,7 +374,7 @@ api.post('/auth', async (ctx, next) => {
                         console.log("[Auth]Create Success: Token Created");
                         const accessToken = token.generateToken({ userId: result.userId });
 
-                        ctx.cookies.set("accessToken", accessToken, { httpOnly: false, maxAge: 1000 * 60 * 60 * 21 });
+                        ctx.cookies.set("accessToken", accessToken, { maxAge: 1000 * 60 * 60 * 21, sameSite: "Strict", overwrite: true });
                         ctx.status = 200;
                     }).catch(err => {
                         console.log(err);
@@ -384,6 +444,13 @@ api.get('/auth', (ctx, next) => {
 // res: 성공 - random QR code data(200) / 에러: Error message(500)
 // 단순 QR code에 들어갈 데이터를 생성하는 것이므로 실패할 수 없음
 api.post('/codes', async (ctx, next) => {
+    if(ctx.request.headers.referer !== process.env.LOGIN_REFERER) {
+        console.log("[QR]Create Failed: Unkown referer");
+        ctx.body = "Bad request.";
+        ctx.status = 400;
+        return;
+    }
+
     // const codeData = crypto.randomBytes(64).toString('hex');
     // const codeData = crypto.randomBytes(32).toString('base64');
     const codeData = uuidv4();
@@ -456,6 +523,13 @@ api.get('/codes/:codeData', async (ctx, next) => {
 // res: 성공 - 로그인이 확인된 경우:OK(200) + Access token(+) / 에러 - Error message(500)
 // GET 방식이 적절하나, QR 인식이 GET 방식인 것을 감안해 PUT 방식으로 호출
 api.put('/codes/:codeData', async (ctx, next) => {
+    if(ctx.request.headers.referer !== process.env.LOGIN_REFERER) {
+        console.log("[QR]Read Failed: Unkown referer");
+        ctx.body = "Bad request.";
+        ctx.status = 400;
+        return;
+    }
+
     const { codeData } = ctx.params;
 
     await model.sequelize.models.QRCodes.findOne({
@@ -480,6 +554,13 @@ api.put('/codes/:codeData', async (ctx, next) => {
 // res: 성공 - OK(200) / 에러 - Error message(500)
 // QR code의 데이터는 생성될 때 DB에 등록되므로 없는 데이터 일수가 없기에 실패할 수 없음
 api.delete('/codes/:codeData', async (ctx, next) => {
+    if(ctx.request.headers.referer !== process.env.LOGIN_REFERER) {
+        console.log("[QR]Delete Failed: Unkown referer");
+        ctx.body = "Bad request.";
+        ctx.status = 400;
+        return;
+    }
+    
     const { codeData } = ctx.params;
 
     await model.sequelize.models.QRCodes.destroy({
